{{{ Finished feature 'Liveset save / restore' The configuration file saves:   * EditList , SampleList and StreamList   * VIMS Bundles   * Customized keybindings
Use veejay -l liveset to load the configuration file.
Use sayVIMS "085:1 liveset;" to save the configuration file.

In GVeejay launcher, press yellow button to load a configuration file
or press the red button to save a configuration file.

more changes:
 GVeejay: new Input Streams Network and File
          disabled speed slider while playing streams

}}}


git-svn-id: svn://code.dyne.org/veejay/trunk@337 eb8d1916-c9e9-0310-b8de-cf0c9472ead5

+++ veejay-current/veejay/libveejay.h
@@ -33,7 +33,7 @@ int veejay_init(veejay_t *info,int w, int h, char *arg, int td);
 
 int veejay_open(veejay_t *info);
 
-int veejay_open_files(veejay_t * info, char **files, int num_files, int fps, int force, int pixfmt);
+int veejay_open_files(veejay_t * info, char **files, int num_files, float fps, int force, int pixfmt);
 
 int veejay_main(veejay_t * info);
